handling project archiving

+++ helpers/projects.js
@@ -35,6 +35,22 @@ export async function joinProject(userId, projectId, offer, userIsMember, setPro
     }
 }
 
+export async function getMembershipsByProject(projectId){
+  let proj = await client.fetch(`
+    [_type == "project" && _id == ${projectId}]{
+      "memberships": *[_type == "ProjectMembership" && references(^._id)]{
+        _id
+      }
+    }
+  `)
+
+  if(proj){
+    return proj.memberships
+  } else {
+    return []
+  }
+}
+
 export async function updateMembership(projectId, _id, quantity, setProcess, router){
   try {
     setProcess({ loading: true, status: "pending" })
@@ -56,13 +72,65 @@ export async function updateMembership(projectId, _id, quantity, setProcess, rou
   }
 }
 
-export function deleteMembership(_id){
-  client.delete(_id)
+export async function archiveProject(projectId, process, setProcess, mode, router){
+  if(!process.loading){
+    setProcess({ loading: true, status: "pending" })
+    try{
+      const archived = await client.patch(projectId)
+        .set({archived: mode === "unArchiving" ? false : true})
+        .commit()
+
+      if(archived){
+        if(mode === "unArchiving"){
+          try{
+            let memberships = await getMembershipsByProject(projectId)
+
+            if(memberships){
+              while(memberships.length !== 0){
+                memberships.map(function({_id}){
+                  client.delete(_id)
+                  let otherMs = memberships.filter((ms) => ms._id !== _id);
+                  memberships = otherMs
+                })
+              }
+
+              if(memberships.length === 0){
+                setProcess({ loading: false, status: "succeed" })
+                router && router.push(`/projects/${projectId}`);
+              }
+            }
+          } catch {
+            setProcess({ loading: false, status: "failed" })
+          }
+        } else {
+          setProcess({ loading: false, status: "succeed" })
+          router && router.push(`/projects/${projectId}`);
+        }
+        return { data: archived }
+      } else {
+        setProcess({ loading: false, status: "failed" })
+      }
+    } catch(error) {
+      setProcess({ loading: false, status: "failed" })
+      return { error }
+    }
+  } else {
+    alert("Processus en cours, merci de patienter !")
+  }
+}
+
+export function deleteMembership(_id, process, setProcess){
+  if(!process.loading){
+    setProcess(function(process){return { ...process, loading: true, status: "pending" }})
+    client.delete(_id)
     .then(function(deleted){
-      console.log({deleted})
+      setProcess(function(process){return { ...process, loading: false, status: "succeed" }})
       location.reload()
     })
-    .catch(function(deletionError){
-      console.log(deletionError)
+    .catch(function(){
+      setProcess(function(process){return { ...process, loading: false, status: "failed" }})
     })
+  } else {
+    alert("Processus en cours, veuillez patienter !")
+  }
 }
